use browser locale in DateFromJulian
